requirements.txt
numpy
scipy
tqdm
tensorflow
pillow
Rtree
trimesh
./flbind 
pyvista
pyacvd
heapdict
libigl
plyfile
scikit-learn

